Update Makefile fallback handling of plan_terraform_module variable call
Update Makefile fallback handling of plan_terraform_module variable call.

Previously this required that there was a file defined within this variable context that is different than the default value for the parameter. This update checks for the existence of that file ahead of run, if it does exist it will honor it else leave it out.

Signed-off-by: jaxom-widby-nttd <jaxom.widby@nttdata.com>